Fix Undo Bar appearance during Google Now use.
Sometimes the Undo Bar can be in a visible state when a Trebuchet user
toggles to use Google Now instead. Since CM Home is used to provide a
transition effect, the Undo Bar should not be visible at that time. Fix
this by setting visibility to GONE when setShowContent(false) is called.

Change-Id: I4f48096500c0dde6dacead3b59ff4c4eb050c30f